+++ astro/src/main.ts
@@ -2,14 +2,20 @@
 import {
   bs,
   HttpConfig,
+  HttpError,
   ServerRequest,
   ServerResponse,
   RouterMatchFunc,
 } from "@bs-core/shell";
 
 import type { AstroIntegration, SSRManifest } from "astro";
+
 import { App } from "astro/app";
 import { pathToFileURL } from "url";
+
+import { performance } from "node:perf_hooks";
+import type { ReadableStream } from "node:stream/web";
+import * as streams from "node:stream/promises";
 
 // Misc consts here
 const ADAPTER_NAME = "@bs-core/astro";
@@ -19,7 +25,6 @@ const HTTP_KEY_FILE = "HTTP_KEY_FILE";
 const HTTP_ENABLE_HTTPS = "HTTP_ENABLE_HTTPS";
 const HTTP_IF = "HTTP_IF";
 const HTTP_PORT = "HTTP_PORT";
-const HTTP_MAINTENANCE_MODE = "HTTP_MAINTENANCE_MODE";
 
 // Module properties here
 
@@ -45,8 +50,6 @@ export type Options = {
   healthcheckPath?: string;
   healthcheckGoodRes?: number;
   healthcheckBadRes?: number;
-
-  maintenanceRoute?: string;
 };
 
 export class WebRequest extends Request {
@@ -54,9 +57,30 @@ export class WebRequest extends Request {
   public res: ServerResponse;
 
   constructor(req: ServerRequest, res: ServerResponse) {
-    // NOTE: WE don't care about headers since they are already on req and res
-    super(req.urlObj, { method: req.method });
+    // Get the headers so we can pass them on
+    const headers = new Headers();
 
+    for (const [name, value] of Object.entries(req.headers)) {
+      // Skip headers with no values
+      if (value === undefined) {
+        continue;
+      }
+
+      // Value can be an array so we need to handle that
+      if (Array.isArray(value)) {
+        // Append each value as the header name
+        for (const item of value) {
+          headers.append(name, item);
+        }
+      } else {
+        headers.append(name, value);
+      }
+    }
+
+    // Now call the super constructor to create the Request
+    super(req.urlObj, { method: req.method, body: req.body, headers });
+
+    // Make sure to tack on the nodeJs req/res
     this.req = req;
     this.res = res;
   }
@@ -65,14 +89,14 @@ export class WebRequest extends Request {
 // Private functions here
 function matcher(_: string): RouterMatchFunc {
   return (url: URL) => {
-    let routeData = _app.match(new Request(url));
+    const routeData = _app.match(new Request(url));
 
     if (routeData === undefined) {
       return false;
     }
 
     // For now we will ignore the params because they are available
-    // in the Astro Request object
+    // in the Web Request object
     // NOTE: We need to return routeData so it can be use with _app.render()
     return {
       params: {},
@@ -85,26 +109,58 @@ async function ssrEndpoint(
   req: ServerRequest,
   res: ServerResponse,
 ): Promise<void> {
-  // Create a WebRequest object to pass to the render
-  let webReq = new WebRequest(req, res);
+  // Create a webRequest object to pass to the render
+  const webReq = new WebRequest(req, res);
+
+  // Measure the time it took to render the page so capture when we started
+  const startedAt = performance.now();
 
   // Now render the page
-  let webRes = await _app.render(webReq, { routeData: req.matchedInfo });
+  const webRes = await _app.render(webReq, {
+    routeData: req.matchedInfo,
+  });
 
-  // First check to see if the response has already been redirected
-  if (res.redirected) {
-    return;
+  // Now figure out how long it took to render and store it in the metrics
+  const latency = Math.round((performance.now() - startedAt) * 1000) / 1000;
+  res.serverTimingsMetrics.push({ name: "ssr", duration: latency });
+
+  // The default status code is always 200. If webRes.status is NOT 200
+  // then it was set by the user so we need to use that status code
+  if (webRes.status !== 200) {
+    res.statusCode = webRes.status;
   }
 
-  // Make sure to get the status code
-  res.statusCode = webRes.status;
+  // Check if the user set any headers on webRes
+  const SET_COOKIE = "set-cookie";
 
-  // This is the easiest way to get the Response body
-  res.body = await webRes.text();
+  for (let [name, value] of webRes.headers.entries()) {
+    // Dont bother setting set-cookie headers, we will do that later
+    if (name === SET_COOKIE) {
+      continue;
+    }
 
-  // Check if Astro set any headers that we should pass on
-  for (let header of webRes.headers.entries()) {
-    res.setHeader(header[0], header[1]);
+    res.setHeader(name, value);
+  }
+
+  // Check if there were any set-cookie headers and if so set them as an array
+  const cookies = webRes.headers.getSetCookie();
+  if (cookies.length > 0) {
+    res.setHeader(SET_COOKIE, cookies);
+  }
+
+  // This is our last chance to set headers so set the server timings header
+  res.setServerTimingHeader();
+
+  // Now check if there is a body in the webRes
+  if (webRes.body !== null) {
+    // NOTE1: pipeline will close the res when it is finished
+    // NOTE2: You need to cast as ReadableStream<Uint8Array> or TS will complain
+    await streams
+      .pipeline(webRes.body as ReadableStream<Uint8Array>, res)
+      .catch((e) => {
+        bs.error("ssrEndpoint had this error during rendering: (%s)", e);
+        throw new HttpError(500, "Internal Server Error");
+      });
   }
 }
 
@@ -162,8 +218,6 @@ export const start = async (
     healthcheckPath: options.healthcheckPath,
     healthcheckGoodRes: options.healthcheckGoodRes,
     healthcheckBadRes: options.healthcheckBadRes,
-    maintenanceRoute: options.maintenanceRoute,
-    startInMaintenanceMode: bs.getConfigBool(HTTP_MAINTENANCE_MODE, false),
   };
 
   if (options.staticFilesPath !== undefined) {
@@ -174,7 +228,7 @@ export const start = async (
     };
   }
 
-  let enableHttps = bs.getConfigBool(HTTP_ENABLE_HTTPS, false);
+  const enableHttps = bs.getConfigBool(HTTP_ENABLE_HTTPS, false);
 
   if (enableHttps) {
     httpConfig.enableHttps = true;
@@ -182,28 +236,28 @@ export const start = async (
     httpConfig.httpsKeyFile = bs.getConfigStr(HTTP_KEY_FILE);
   }
 
-  let networkIf = bs.getConfigStr(HTTP_IF, "127.0.0.1");
-  let networkPort = bs.getConfigNum(HTTP_PORT, 8080);
+  const networkIf = bs.getConfigStr(HTTP_IF, "127.0.0.1");
+  const networkPort = bs.getConfigNum(HTTP_PORT, 8080);
 
   // Create the HTTP server
-  // NOTE: Don't start it until we are finished setting everything up
-  let httpServer = await bs.addHttpServer(
+  const httpServer = await bs.addHttpServer(
     networkIf,
     networkPort,
     httpConfig,
-    false,
+    false, // NOTE: Don't start until we are finished setting everything up
   );
 
   // Call setupEntryPoint here in case you want to setup any default
   // middleware for the SSR endpoint
   if (options.setupEntryPoint !== undefined) {
     // NOTE: We expect an exported function named "setup"
-    let { setup } = await import(pathToFileURL(options.setupEntryPoint).href);
+    const { setup } = await import(pathToFileURL(options.setupEntryPoint).href);
     await setup();
   }
 
-  // Add the main SSR route - NOTE: the path is not important
-  httpServer.ssrRouter.get("/", ssrEndpoint, {
+  // Add the main SSR route - NOTE: the path is not important since the
+  // matcher will decide if there is a matching page
+  httpServer.ssrRouter.all("/", ssrEndpoint, {
     generateMatcher: matcher,
     etag: true,
   });
@@ -211,5 +265,5 @@ export const start = async (
   // Start the http server now!
   await httpServer.start();
 
-  bs.startupMsg("Astro adapter is primed - party on!!");
+  bs.startupMsg("Astro adapter is primed - party on dudes!!");
 };